Update from DR1 (Jacky)
Add new file "quantification.py" in "directive" folder, and move the class "qestimate" from "benchmark.py" to "quantification.py"
